// packages/desktop/views/onboarding/onboarding-route.enum.ts
export enum OnboardingRoute {
    Welcome = 'welcome',
    NetworkSetup = 'networkSetup',
    ChooseOnboardingFlow = 'chooseOnboardingFlow',
    CreateProfile = 'createProfile',
    RestoreProfile = 'restoreProfile',
    CompleteOnboarding = 'completeOnboarding',
    Congratulations = 'congratulations',
    LedgerSetup = 'ledgerSetup',
    StrongholdSetup = 'strongholdSetup',
    ProfileBackup = 'profileBackup',
    ProfileRecovery = 'profileRecovery',
    ProfileSetup = 'profileSetup',
    StorageProtectionSetup = 'storageProtectionSetup',
    ShimmerClaiming = 'shimmerClaiming',
}